Last update before presentation to M. Rieder

+++ clockDesign/Controller/digitalclock.h
@@ -9,16 +9,16 @@ class DigitalClock
 {
 public:
     DigitalClock(Controller* host);
-    void set_0(int digit);
-    void set_1(int digit);
-    void set_2(int digit);
-    void set_3(int digit);
-    void set_4(int digit);
-    void set_5(int digit);
-    void set_6(int digit);
-    void set_7(int digit);
-    void set_8(int digit);
-    void set_9(int digit);
+    void set_0(double digit);
+    void set_1(double digit);
+    void set_2(double digit);
+    void set_3(double digit);
+    void set_4(double digit);
+    void set_5(double digit);
+    void set_6(double digit);
+    void set_7(double digit);
+    void set_8(double digit);
+    void set_9(double digit);
 
 private:
     Controller* host;
